BLE: Move advertising data types in their own file.
